server/requirements.txt
Flask>=2.0
flask-cors>=3.0
pandas>=2.0
numpy>=1.25
plotly>=5.0
yfinance>=0.2
flask-session>=0.4.0
flask-bcrypt>=1.0.0
redis>=4.0.0
python-dotenv>=1.0.0
requests>=2.0
